+++ shortcut.ps1
@@ -6,11 +6,11 @@ function Set-ShortcutRunAsAdministrator {
     .DESCRIPTION
         Sets the "Run as administrator" flag on the specified shortcut.
 
-    .PARAMETER ShortcutLocation
+    .PARAMETER Shortcut
         The path to the shortcut to set the "Run as administrator" flag on.
 
     .EXAMPLE
-        Set-ShortcutRunAsAdministrator -ShortcutLocation "C:\Users\UserName\Desktop\MyShortcut.lnk"
+        Set-ShortcutRunAsAdministrator -Shortcut "C:\Users\UserName\Desktop\MyShortcut.lnk"
 
     .NOTES
         This function is based on the answer at https://stackoverflow.com/a/29002207/2562544.
@@ -18,19 +18,18 @@ function Set-ShortcutRunAsAdministrator {
     [CmdletBinding(SupportsShouldProcess)]
     param (
         [Parameter(Mandatory = $true)]
-        [Alias("Location")]
-        [string] $ShortcutLocation
+        [string] $Shortcut
     )
 
-    $bytes = [System.IO.File]::ReadAllBytes($ShortcutLocation)
+    $bytes = [System.IO.File]::ReadAllBytes($Shortcut)
     $bytes[0x15] = $bytes[0x15] -bor 0x20 #set byte 21 (0x15) bit 6 (0x20) ON
 
-    if ($PSCmdlet.ShouldProcess($ShortcutLocation, "Set 'Run as administrator' flag")) {
-        [System.IO.File]::WriteAllBytes($ShortcutLocation, $bytes)
+    if ($PSCmdlet.ShouldProcess($Shortcut, "Set 'Run as administrator' flag")) {
+        [System.IO.File]::WriteAllBytes($Shortcut, $bytes)
     }
 }
 
-function Get-ShortcutIconLocation {
+function Get-ShortcutIcon {
     <#
     .SYNOPSIS
         Returns the icon location of a shortcut.
@@ -38,23 +37,23 @@ function Get-ShortcutIconLocation {
     .DESCRIPTION
         Returns the icon location of the specified shortcut.
 
-    .PARAMETER Location
+    .PARAMETER Shortcut
         The path to the shortcut to read the icon location from.
 
     .OUTPUTS
         string - The icon location of the shortcut.
 
     .EXAMPLE
-        Get-ShortcutIconLocation -Location "C:\Users\UserName\Desktop\MyShortcut.lnk"
+        Get-ShortcutIcon -Shortcut "C:\Users\UserName\Desktop\MyShortcut.lnk"
     #>
     [CmdletBinding()]
     param (
         [Parameter(Mandatory = $true)]
-        [string] $Location
+        [string] $Shortcut
     )
 
     $wshShell = New-Object -ComObject WScript.Shell
-    $shortcut = $wshShell.CreateShortcut($Location)
+    $shortcutObject = $wshShell.CreateShortcut($Shortcut)
 
-    return $shortcut.IconLocation
+    return $shortcutObject.IconLocation
 }